Move nav component functionality into app component
